Power Pellets, Point Boosts, Level Advance
Reworked power pellet system, added point boost fruits, and added a
level advance system (but all the levels are still the same like
Pacman).

Once I add sound to the game, it will be mostly finished.
Might try and integrate the level editor (which can also be found on my
Github) into this app, but for now this is about it.

Known issues:
- white flash between levels (epilepsy warning probably)